cyber-service/service/requirements.txt
# Web framework
Flask==2.2.2

# Environment variable management
python-dotenv

# MongoDB interaction
pymongo[srv]==4.2.* 
mongoengine
mongoengine_softdelete

# API documentation and REST support
flasgger
flask_restx

# CORS handling
flask-cors

# Serialization and validation
marshmallow

# Certificates
certifi

# Ensure compatibility with Flask
Werkzeug==2.2.2

# JWT handling
flask-jwt-extended==4.4.4

# Transformers library for NLP tasks
transformers

# Add TensorFlow
# tensorflow==2.13.0

# Timezone support
pytz

gunicorn

Flask-CORS

openai


google-cloud-scheduler

marshmallow==3.14.0

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-xdist==3.3.1
mongomock==4.1.2
flask-testing==0.8.1